feat: I put functions and types in the global file
